增加了csv读写的方法，可以从本地csv文件恢复原有数据了；增加了一些注释

+++ app/src/main/java/com/example/pc/newble/TheUtils/FileUtils.java
@@ -2,8 +2,10 @@ package com.example.pc.newble.TheUtils;
 
 
 import android.os.Environment;
+import android.support.design.widget.Snackbar;
 import android.util.Log;
 
+import java.io.BufferedWriter;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.Vector;
@@ -26,7 +28,9 @@ public class FileUtils {
 
     private static final String TAG = "FileUtility";
 
-    // 将字符串写入到文本文件中
+    /**
+     * 将字符串写入到文本文件中
+     */
     public static void writeTxtToFile(String strcontent, String filePath, String fileName) {
         //生成文件夹之后，再生成文件，不然会出错
         makeFilePath(filePath, fileName);
@@ -118,8 +122,8 @@ public class FileUtils {
     }
 
 
-    /** 此函数仅给 RetrieveData.java 下的读取人体电压用
-     * 对于其他函数请勿使用，因为有可能造成数据丢失
+    /**
+     * 此函数仅给 RetrieveData.java 下的读取人体电压用
      */
     public static Vector<Double> readVoltageFromFile(String file) {
 
@@ -134,8 +138,6 @@ public class FileUtils {
             Log.e(TAG, "readVoltageFromFile: 读取行内内容成功：(第一行)" + str);
             str = in.readLine();
             Log.e(TAG, "readVoltageFromFile: 读取行内内容成功：(第二行)" + str);
-
-            // TODO 可能会flush掉第一行的内容，根据后续设定再改
 
             while ((str = in.readLine()) != null) {
                 Log.e("Retrieve.this QQQQQ", "readVoltageFromFile: "+temp + "   " + savedDoubles.size());
@@ -185,12 +187,11 @@ public class FileUtils {
                     savedStrings.add(str);
                 }
             }
-            //最后一行的处理：似乎永远是空格？
-            //System.out.println(str);
+
         } catch (java.io.IOException e) {
 
         }
-        Log.e(TAG, "readVoltageFromFile: 读取函数的读取结束" );
+        Log.d(TAG, "readVoltageFromFile: 读取函数的读取结束" );
         return savedStrings;
     }
 
@@ -200,8 +201,12 @@ public class FileUtils {
      */
     public static void addStringToFile(String fileName, String content) {
         try {
-            // TODO 如果不存在的话，新建一个
-
+            // 检查是否存在这样的文件，不存在则报错
+            File f = new File(fileName);
+            if (f.exists() == false){
+                Log.e(TAG, "addStringToFile: File does not exist" );
+                return;
+            }
 
             // 打开一个写文件器，构造函数中的第二个参数true表示以追加形式写文件
             FileWriter writer = new FileWriter(fileName, true);
@@ -232,59 +237,17 @@ public class FileUtils {
 
 
     /**
-     * 根据 /bletest/ 目录下的内容，自动生成更新后的 DataList.txt 的函数。
-     * 注：本函数的使用位置有些奇怪，因为 MainActivity 里的函数有些叠床架屋，故放到了调用 ChooseHistActivity 时
-     *    触发。同时，这样做也带来了一个吊诡的好处 —— 不害怕用户瞎改 DataList.txt 了……
-     * WARNING: 本函数高度特化
+     * 向 csv 文件写入的函数。
      * */
-    public static void updateDataList(){
-        String path = getSDCardPath() + "/bletest/DataList.txt";
-        File file = new File(path);
-        try {
-            if (file.exists() == false) {
-                // 如果不存在这个文件，就新建一个
-                file.createNewFile();
-            } else {
-                // 如果存在，就清空
-                FileWriter fileWriter =new FileWriter(file);
-                fileWriter.write("");
-                fileWriter.flush();
-                fileWriter.close();
-            }
-
-            // 得到目录下所有以"20"开头的字符，并将它们写入 DataList.txt
-            // （显然我们都不指望这个 app 活到二十二世纪）
-            Vector<String> allName = getFilesAllName(getSDCardPath() + "/bletest/");
-            Log.e(TAG, "updateDataList: " + allName.get(0) );
-            Vector<String> histData = new Vector<>();
-            for (String item : allName){
-          //      Log.e(TAG, "updateDataList: "+item );
-            //    Log.e(TAG, "updateDataList: "+item.substring(item.length()-18, item.length()-0) );
-              //  Log.e(TAG, "updateDataList: " );
-                if (item.length() >= 44){
-                    Log.e(TAG, "updateDataList: 已经添加了一项HistData" );
-                    histData.add(item.substring(item.length()-18, item.length()-0));
-                }
-
-            }
-            Log.e(TAG, "updateDataList: HISTDATA"+histData.get(0) );
-            for (String string : histData){
-                // 向文件写入换行符
-                addStringToFile(path, "\n");
-                // 向文件写入传入函数的实参
-                addStringToFile(path, string);
-            }
-
-
-        } catch (Exception e){
-            e.printStackTrace();
+    public static void addLineToCsvFile(String path, Vector<String> inputs) {
+        String string = new String();
+        for (int i=0; i<inputs.size(); i++) {
+            string = string + "," + inputs.get(i);
         }
-
+        string = string + "\n";
+        addStringToFile(path, string);
     }
+
 
 }
 
-      /*  作者：Sky*ZST
-        来源：CSDN
-        原文：https://blog.csdn.net/u012246458/article/details/83063112
-        版权声明：本文为博主原创文章，转载请附上博文链接！*/
